pyproject.toml
[project]
name = "marketpilot"
version = "0.1.0"
description = "AI-powered trading analysis platform with multi-agent architecture"
requires-python = ">=3.12"
license = { text = "MIT" }

dependencies = [
    # --- Core Web Framework ---
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.34.0",
    "pydantic>=2.10.0",
    "pydantic-settings>=2.7.0",
    "python-dotenv>=1.0.0",
    "httpx>=0.28.0",

    # --- Data Sources ---
    "yfinance>=0.2.50",
    "edgartools>=3.0.0",
    "finnhub-python>=2.4.0",
    "alpaca-py>=0.38.0",
    "tradingview-screener>=2.7.0",
    "tradingview-ta>=3.3.0",

    # --- Economic Data ---
    "fredapi>=0.5.0",

    # --- Databases ---
    "psycopg2-binary>=2.9.0",
    "redis>=5.2.0",
    "chromadb>=0.6.0",

    # --- LangChain / AI ---
    "langchain-core>=0.3.0",
    "langgraph>=0.2.0",
    "langchain-google-genai>=2.0.0",
    "langsmith>=0.2.0",

    # --- Task Queue ---
    "celery[redis]>=5.4.0",

    # --- Authentication ---
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.0",
    "email-validator>=2.0.0",

    # --- Technical Analysis ---
    "ta>=0.11.0",
    "py-vollib>=1.0.1",
    "mibian>=0.1.3",
    "greeks-package>=1.0.0",
    "finta>=1.3",
    "scipy>=1.14.0",
    "numpy>=2.0.0",
    "pandas>=2.2.0",

    # --- Logging & Visualization ---
    "structlog>=24.0.0",
    "plotly>=5.24.0",

    # --- Timezone ---
    "pytz>=2024.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.25.0",
    "pytest-cov>=6.0.0",
    "ruff>=0.9.0",
    "httpx>=0.28.0",
]
ml = [
    "torch>=2.5.0",
    "torchvision>=0.20.0",
    "ultralytics>=8.3.0",
    "scikit-learn>=1.6.0",
    "mlflow>=2.19.0",
    "pyts>=0.14.0",
]
backtest = [
    "vectorbt>=0.26.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["backend/app"]

[tool.ruff]
target-version = "py312"
line-length = 120

[tool.ruff.lint]
select = ["E", "W", "F", "I", "N", "UP"]

[tool.pytest.ini_options]
testpaths = ["backend/tests"]
asyncio_mode = "auto"
